Improved code (WIP)
